fix(ios): split watch notify normalization helpers
